Handle remark behaviour of definitions
