+++ tests/fixtures.md
@@ -3,26 +3,28 @@ not changelog test
 This is the input Markdown test,
 then below add the expected output.
 
-- a [reference][]
-- b [0.0.1]
+- a [link]
+- b [0.0.1][]
 * c [1.0.0]
 
 [unused reference]: https://example.com
 [0.0.1]: https://example.com
-[reference]: https://example.com
+[link]: https://example.com
 [1.0.0]: https://example.com
 .
 This is the input Markdown test,
 then below add the expected output.
 
-- a [reference]
-- b [0.0.1]
+- a [link]
+- b [0.0.1][]
 
 * c [1.0.0]
 
 [0.0.1]: https://example.com
+
 [1.0.0]: https://example.com
-[reference]: https://example.com
+
+[link]: https://example.com
 .
 
 correct changelog test
@@ -160,7 +162,6 @@ unsorted changelog test
 [0.0.1]: https://example.com
 .
 
-
 extra links changelog test
 .
 # Changelog
@@ -169,16 +170,18 @@ extra links changelog test
 
 ### Added
 
-- Commit [JIRA-2]
+- Commit [JIRA-10]
 
-## [1.1.0] - 2025-01-01
+## [1.1.0][] - 2025-01-01
 
 ### Added
 
+- Commit [JIRA-2]
 - Commit [JIRA-1]
 
-[JIRA-2]: https://example.com
-[JIRA-1]: https://example.com
+[JIRA-10]: https://example.com "ticket title"
+[JIRA-2]: https://example.com "ticket title"
+[JIRA-1]: https://example.com "ticket title"
 
 [2.10.0]: https://example.com
 [10.1.0]: https://example.com
@@ -189,18 +192,22 @@ extra links changelog test
 
 ### Added
 
-- Commit [JIRA-2]
+- Commit [JIRA-10]
 
-## [1.1.0] - 2025-01-01
+## [1.1.0][] - 2025-01-01
 
 ### Added
 
+- Commit [JIRA-2]
 - Commit [JIRA-1]
 
 [10.1.0]: https://example.com
 
 [2.10.0]: https://example.com
 
-[jira-1]: https://example.com
-[jira-2]: https://example.com
+[JIRA-1]: https://example.com "ticket title"
+
+[JIRA-10]: https://example.com "ticket title"
+
+[JIRA-2]: https://example.com "ticket title"
 .
